requirements.txt
Flask==3.0.2
charm-crypto==0.50
Werkzeug==3.0.1 